Update :csrf_attack to "csrf_attack"
The `Ueberauth.Failure.Error` struct defines a type specification that
says that the `message_key` field should be a binary, however in the case of
CSRF attacks the `Ueberauth.Strategy` sets the error's `message_key` to
`:csrf_attack`. Whilst this does not cause problems in the way this
library works, it causes discrepancies we the way client code has to
handle errors. Ueberauth extensions such as `Ueberauth.Google` follow
the correct type specification and use binaries for `message_key`, so
client code needs to be able to pattern match against different types
which is not ideal.