Fix default port selection for the callback_url (#181) (#181)
Cowboy will automatically default to port 80 when no port has been
given, but Bandit does not. Bandit will always default to the
underlying port, which makes it hard to know if there was a port
specified or if a port was assumed.

This change ignores cowboy or bandit's assumption around port number
and parses it from the header directly.